Move testutil in pkg/kube
