+++ pkg/kube/lists/containers/containers_test.go
@@ -3,7 +3,7 @@ package containers
 import (
 	"testing"
 
-	"github.com/stakater/Konfigurator/pkg/testutil"
+	"github.com/stakater/Konfigurator/pkg/kube/testutil"
 	corev1 "k8s.io/api/core/v1"
 )
