+++ pkg/kube/mounts/manager_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stakater/Konfigurator/pkg/apis/konfigurator/v1alpha1"
 	"github.com/stakater/Konfigurator/pkg/kube/lists/containers"
 	"github.com/stakater/Konfigurator/pkg/kube/lists/volumes"
-	"github.com/stakater/Konfigurator/pkg/testutil"
+	"github.com/stakater/Konfigurator/pkg/kube/testutil"
 )
 
 func TestMountVolumes(t *testing.T) {
